refine ui for message screen

+++ app/src/main/java/com/xctbtx/cleanarchitectsample/data/user/dto/UserDto.kt
@@ -4,7 +4,7 @@ data class UserDto(
     val id: String = "",
     val name: String = "",
     val username: String = "",
-    val address: String= "",
+    val address: String = "",
     val avatar: String = "",
-    val dob: Long = 0
+    val dob: Int = 0
 )
